feat: add run command
Adds support for the "juju run" command. Don't support background mode
or show-operation for now (likely not needed for integration testing).

This PR adds a dependency on PyYAML, which is kind of inevitable at
some point. All params are passed as YAML, to ensure they round trip
with the correct types (and to avoid ad-hoc foo.bar=baz serialization
on the command line).

Fixes #13