feat: add ssh command
This PR adds support for "juju ssh":

    def ssh(
        self,
        *command: str,
        machine: int | None = None,
        unit: str | None = None,
        container: str | None = None,
        host_key_checks: bool = True,
        ssh_options: list[str] | None = None,
        user: str | None = None,
    ) -> str:

    >>> juju.ssh('echo foo', unit='snappass-test/0')
    'foo\n'

The arguments aren't in quite the same order as the CLI, but rather
"machine" and "unit" are done as two separate keyword arguments, for
better types and for consistency with the `exec` method.

The "user" is also a separate argument, not stuffed into the target as
on the CLI.

One decision point was the host_key_checks argument. I lean towards
what I have here, but we could also do `no_host_key_checks: bool = False`,
which is more 1:1 with the CLI. It just felt weird to have an inverted
"no_foo" parameter when we can just invert the default in Python.

Fixes #14