feat: add Juju 2.9 compatibility (#193)
Adds compatibility with Juju 2.9. The API remains the same as the
regular Jubilant (meaning the Juju 3.x API) with the code translating to
Juju 2.9 CLI calls where required, if the Juju version is 2.9.

This is essentially the same (and replaces)
https://github.com/tonyandrewmeyer/jubilant-backports, but because it's
in a branch of the actual Jubilant code rather than a separate package,
it's much simpler because instead of lots of subclassing there are more
surgical changes.

There is one new public name, `ExecTask`, because in Juju 2.9 `exec` and
`run-action` have quite different return types.

The `Juju` class loads the CLI version at instantiation time, which does
mean there is one extra subprocess call per `Juju` object created, but
that seems ok for this backports package (it could be moved to a lazy
property if required). We are expecting to [load the version in a future
release of the main
package](https://github.com/canonical/jubilant/issues/178), but if that
is `.version` it would not clash with this.

The major differences between 2.9 and 3.x are:
* 2.9 does not have secrets. We let the Juju "not a juju command" error
surface naturally.
* `base` arguments are translated to `series` in `bootstrap()`,
`deploy()`, and `refresh()`.
* `exec()` has a different return type, which means that the signature
changes slightly, to `Task | ExecTask`.
* `run` works somewhat differently in Juju 2.9.

The same status classes are used for 2.9 and 3.x. They are mostly the
same, and where there are differences the fields will be empty in 2.9 or
3.x. The most significant difference is in `AppStatus`, where
`relations` is a `dict[str, list[AppStatusRelation]]` but in 2.9 none of
the fields of the `AppStatusRelation`s are filled out. There are two new
public names: `MeterStatus` and `BranchStatus`.

The existing unit tests remain generally unchanged, except that the unit
tests are set to behave as if they are running under Juju 3. A new set
of unit tests that explicitly test both 2.9 and 3.x where the behaviour
differs is added.

The existing integration tests remain generally unchanged. A new set of
integration tests that explicitly test both 2.9 and 3.x where the
behaviour differs is added.

The intention is **not** to release this as a new `jubilant` package
release. Instead a follow-up PR will adjust `pyproject.toml` and some
documentation to release this as a new `jubilant-backports` package
release. We also intend to only provide bug and security fixes for the
2.9 support: most particularly, if any new features are added to the
main Jubilant releases, they will not be backported to the
`jubilant-backports` branch.